v8 bug fixes - available memory now returns correct amount, removed average cthroughput and execution time calculation

+++ changes/v8changes/src/heap/heap.h
@@ -1052,7 +1052,7 @@ class Heap {
   // ===========================================================================
   
   //setter for max_old_generation_size_ 
-  void setMaxOldGenerationSize(int max_old_space);
+  void setMaxOldGenerationSize(long long max_old_space);
   long long getMaxOldGenerationSize(){ return (long long)max_old_generation_size_; }
 
   // Configure heap size in MB before setup. Return false if the heap has been
